requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
python-decouple
python-slugify
